% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{plotNucleosomes}
\alias{plotNucleosomes}
\title{Generate a graph of nucleosome position}
\usage{
plotNucleosomes(nucleosomesPosition, reads, xlab = "position",
  ylab = "coverage")
}
\arguments{
\item{nucleosomesPosition}{a \code{list} or a \code{vector} of
\code{numeric}, the nucleosome positions for one or
multiples results obtained using the same reads. TODO}

\item{reads}{an \code{IRanges} containing all the reads}

\item{xlab}{a \code{character} string containing the label of the x-axis}

\item{ylab}{a \code{character} string containing the label of the y-axis}
}
\value{
a graph containing the nucleosome position and the read coverage
}
\description{
Generate a graph for
a list or a vector of nucleosome positions. TODO
}
\examples{

## Generate a synthetic sample with 10 well-positioned nucleosomes, 2 fuzzy
## nucleosomes and 2 deleted nucleosomes using a Student distribution
## with a variance of 10 for the well-positioned nucleosomes,
## a variance of 20 for the fuzzy nucleosomes
library(nucleoSim)
nucleosomeSample <- syntheticNucReadsFromDist(wp.num=10, wp.del=2,
wp.var=10, fuz.num=2, fuz.var=20, max.cover=100, dist="Student",
nuc.len=147, len.var=12, read.len=45, lin.len=20, rnd.seed=155, offset=100)

dataIP <-nucleosomeSample$dataIP

forwardReads <- dataIP[dataIP$strand == "+",]$start
reverseReads <- dataIP[dataIP$strand == "-",]$end

result <- rjmcmc(startPosForwardReads = forwardReads,
         startPosReverseReads = reverseReads,
         nbrIterations = 4000, lambda = 2, kMax = 30,
         minInterval = 146, maxInterval = 292, minReads = 5, vSeed = 10213)

reads <-IRanges(start = dataIP$start, end=dataIP$end)

## Create graph using the synthetic map
plotNucleosomes(nucleosomesPosition = result$mu, reads = reads)

}
\author{
Astrid Deschenes
}

