% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rjmcmcNucleosomesBeforeAndAfterPostTreatment.R
\name{print.rjmcmcNucleosomesBeforeAndAfterPostTreatment}
\alias{print.rjmcmcNucleosomesBeforeAndAfterPostTreatment}
\title{Formated output of predicted nucleosomes}
\usage{
\method{print}{rjmcmcNucleosomesBeforeAndAfterPostTreatment}(x, ...)
}
\arguments{
\item{x}{the output object from \code{rjmcmcCHR}
function to be printed}

\item{\ldots}{arguments passed to or from other methods}
}
\value{
an object of class
\code{rjmcmcNucleosomesBeforeAndAfterPostTreatment}
}
\description{
Generated a formated output of a list marked as
an \code{rjmcmcNucleosomesBeforeAndAfterPostTreatment} class
}
\examples{

## TODO

}
\author{
Astrid Deschenes
}

