% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{validatePlotNucleosomesParameters}
\alias{validatePlotNucleosomesParameters}
\title{Parameters validation for the \code{\link{plotNucleosomes}} function}
\usage{
validatePlotNucleosomesParameters(nucleosomePositions, reads, xlab, ylab, names)
}
\arguments{
\item{nucleosomePositions}{a \code{list} or a \code{vector} of
\code{numeric}, the nucleosome positions for one or
multiples predictions are obtained using the same reads. In presence of
only one prediction (with multiples nucleosome positions), a \code{vector}
is used. In presence of more thant one predictions (as example, before and
after post-treatment or results from different software), a \code{list} with
one entry per prediction is used.}

\item{reads}{an \code{IRanges} containing all the reads.}

\item{xlab}{a \code{character} string containing the label of the x-axis.}

\item{ylab}{a \code{character} string containing the label of the y-axis.}

\item{names}{a \code{vector} of a \code{character} string containing the
label of each prediction set. The \code{vector} must be the same length of
the \code{nucleosomePositions} \code{list} or 1 in presence of a
\code{vector}.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Validation of all parameters needed by the public
\code{\link{plotNucleosomes}} function.
}
\examples{

## Create a IRanges object with 2 reads
reads <- IRanges(start=c(950, 969), end=c(1020, 1022))

## Create a vector containing nucleosome positions
nucleosomes <- c(1001)

## The function returns 0 when all parameters are valid
RJMCMCNucleosomes:::validatePlotNucleosomesParameters(nucleosomePositions =
nucleosomes, reads = reads, xlab = "position", ylab = "coverage",
names = c("test"))

## The function raises an error when at least one paramater is not valid
#\dontrun{RJMCMCNucleosomes:::validatePlotNucleosomesParameters(
#nucleosomePositions = c("hi"), reads = reads, xlab = "position",
#ylab = "coverage", names = c("test"))}

#\dontrun{RJMCMCNucleosomes:::validatePlotNucleosomesParameters(
#nucleosomePositions = nucleosomes, reads = reads, xlab = "position",
#ylab = "coverage", names = c("test_one", "test_false"))}

}
\author{
Astrid Deschenes
}
\keyword{internal}

