% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{plotNucleosomes}
\alias{plotNucleosomes}
\title{Generate a graph of nucleosome positions with read coverage}
\usage{
plotNucleosomes(nucleosomePositions, reads, xlab = "position",
  ylab = "coverage", names = NULL)
}
\arguments{
\item{nucleosomePositions}{a \code{list} or a \code{vector} of
\code{numeric}, the nucleosome positions for one or
multiples predictions are obtained using the same reads. In presence of
only one prediction (with multiples nucleosome positions), a \code{vector}
is used. In presence of more thant one predictions (as example, before and
after post-treatment or results from different software), a \code{list} with
one entry per prediction is used.}

\item{reads}{an \code{IRanges} containing all the reads.}

\item{xlab}{a \code{character} string containing the label of the x-axis.}

\item{ylab}{a \code{character} string containing the label of the y-axis.}

\item{names}{a \code{vector} of a \code{character} string containing the
label of each prediction set. The \code{vector} must be the same length of
the \code{nucleosomePositions} \code{list} or 1 in presence of a
\code{vector}. When \code{NULL}, the name of the elements of the \code{list}
are used or the string "Nucleosome" for a \code{vector} are used.
Default: \code{NULL}.}
}
\value{
a graph containing the nucleosome position and the read coverage
}
\description{
Generate a graph for
a \code{list} or a \code{vector} of nucleosome positions. In presence of
only one prediction (with multiples nucleosome positions), a \code{vector}
is used. In presence of more thant one predictions (as example, before and
after post-treatment or results from different software), a \code{list} with
one entry per prediction is used. All predictions must have been obtained
using the same reads.
}
\examples{

## Generate a synthetic sample with 10 well-positioned nucleosomes, 2 fuzzy
## nucleosomes and 2 deleted nucleosomes using a Student distribution
## with a variance of 10 for the well-positioned nucleosomes,
## a variance of 20 for the fuzzy nucleosomes
library(nucleoSim)
nucleosomeSample <- syntheticNucReadsFromDist(wp.num=10, wp.del=2,
wp.var=10, fuz.num=2, fuz.var=20, max.cover=100, dist="Student",
nuc.len=147, len.var=12, read.len=45, lin.len=20, rnd.seed=155, offset=100)

dataIP <-nucleosomeSample$dataIP

forwardReads <- dataIP[dataIP$strand == "+",]$start
reverseReads <- dataIP[dataIP$strand == "-",]$end

result <- rjmcmc(startPosForwardReads = forwardReads,
            startPosReverseReads = reverseReads,
            nbrIterations = 4000, lambda = 2, kMax = 30,
            minInterval = 146, maxInterval = 292, minReads = 5, vSeed = 10213)

reads <-IRanges(start = dataIP$start, end=dataIP$end)

## Create graph using the synthetic map
plotNucleosomes(nucleosomePositions = result$mu, reads = reads)

}
\author{
Astrid Deschenes
}

